requirement.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses==0.8
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
Markdown==3.3.6
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.62
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
six==1.16.0
sk-video==1.1.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
typing_extensions==4.1.0
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
